[Update] class and assignment

+++ src/helpers/deleteOne.js
@@ -1,4 +1,10 @@
+import {getElementByField} from './getElementByField.js';
+
 export async function deleteOne({model, id}) {
+  const element = await getElementByField({model, field: 'id', value: id});
+
+  if (!element) return null;
+
   return await model.destroy({
     where: {
       id: id,
